requirements.txt
flask
flask-cors
newsapi-python
scikit-learn
nltk
pandas
numpy
vaderSentiment